+++ Server/sources/protocols/ai/init_action_handlers.c
@@ -7,7 +7,7 @@
 
 #include "ai/handlers/init_action_handlers.h"
 
-act_handler_t *init_action_handler(char *command, uint16_t limit_cycles,
+act_handler_t *init_action_handler(char *command, int limit_cycles,
     uint16_t args_nbr, action_handler_fcnt handler)
 {
     act_handler_t *action = malloc(sizeof(act_handler_t));
